Refactor Api Gateway smoke tests
